git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1174 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/num/random.inl
@@ -150,7 +150,7 @@ void RandomMT19937::seed(ForwardIterator first, ForwardIterator last)
 
 	seed(19650218);
 
-	const size_t keySize = std::distance(first, last);
+	const size_t keySize = static_cast<size_t>(std::distance(first, last));
 	LASS_ASSERT(keySize < num::NumTraits<TValue>::max);
 
 	TValue i = 1;
